Revert "Update UserTable to allow single selection"
