+++ kolibri/core/assets/src/core-app/apiSpec.js
@@ -74,7 +74,6 @@ import CoreMenuDivider from '../views/CoreMenu/CoreMenuDivider';
 import CoreMenuOption from '../views/CoreMenu/CoreMenuOption';
 import heartbeat from '../heartbeat';
 import CoreTable from '../views/CoreTable';
-import UserTable from '../views/UserTable';
 import CoachContentLabel from '../views/CoachContentLabel';
 import PrivacyInfoModal from '../views/PrivacyInfoModal';
 import UserTypeDisplay from '../views/UserTypeDisplay';
@@ -174,7 +173,6 @@ export default {
       CoreMenuDivider,
       CoreMenuOption,
       CoreTable,
-      UserTable,
       CoreInfoIcon,
       InteractionList,
       ExamReport,
