+++ kolibri/plugins/setup_wizard/assets/src/views/importLODUsers/MultipleUsers.vue
@@ -49,7 +49,7 @@
   import commonSyncElements from 'kolibri.coreVue.mixins.commonSyncElements';
   import PaginatedListContainer from 'kolibri.coreVue.components.PaginatedListContainer';
   import { TaskStatuses } from 'kolibri.utils.syncTaskUtils';
-  import UserTable from 'kolibri.coreVue.components.UserTable';
+  import UserTable from '../../../../../facility/assets/src/views/UserTable.vue';
   import OnboardingForm from '../onboarding-forms/OnboardingForm';
   import { SoudQueue } from '../../constants';
